Move utility functions into own helper class
