+++ conditions/type/type_interface.php
@@ -58,32 +58,6 @@ interface type_interface
 	public function get_group_rules($type);
 
 	/**
-	 * Get user's group ids
-	 *
-	 * @param array $user_id_ary An array of user ids to check
-	 * @return array An array of usergroup ids each user belongs to
-	 * @access public
-	 */
-	public function get_users_groups($user_id_ary);
-
-	/**
-	 * Get users that should not have their default status changed
-	 *
-	 * @return array An array of user ids
-	 * @access public
-	 */
-	public function get_default_exempt_users();
-
-	/**
-	 * Prepare user ids for querying
-	 *
-	 * @param mixed $user_ids User id(s) expected as int or array
-	 * @return array An array of user id(s)
-	 * @access public
-	 */
-	public function prepare_users_for_query($user_ids);
-
-	/**
 	 * Add user(s) to group
 	 *
 	 * @param array $user_id_ary     User(s) to add to group
