+++ tests/conditions/warnings_test.php
@@ -21,8 +21,8 @@ class warnings_test extends type_test_case
 	{
 		return new \phpbb\autogroups\conditions\type\warnings(
 			$this->phpbb_container,
-			$this->config,
 			$this->db,
+			$this->helper,
 			$this->user,
 			'phpbb_autogroups_rules',
 			'phpbb_autogroups_types',
